AI +QA -Bank/requirements.txt
Flask==3.1.0
selenium==4.30.0
pytest==8.3.5
allure-pytest==2.13.5
requests==2.32.3
